Sequence counter as translation unit, added namespace
