+++ sakovskaia.maiia/P1/main.cpp
@@ -1,33 +1,11 @@
 #include <iostream>
-#include <limits>
 #include <stdexcept>
-
-struct SequenceCounter
-{
-  SequenceCounter():
-    count_(0)
-  {}
-
-  void count(int num)
-  {
-    size_t max_size = std::numeric_limits< size_t >::max();
-    if (count_ == max_size)
-    {
-      throw std::logic_error("Sequence is too long");
-    }
-    ++count_;
-  }
-  size_t get_result() const
-  {
-    return count_;
-  }
-private:
-  size_t count_;
-};
+#include "sequence_counter.hpp"
 
 int main()
 {
   int number = 0;
+  using namespace sakovskaia;
   SequenceCounter counter;
   do
   {
@@ -41,7 +19,7 @@ int main()
     {
       try
       {
-        counter.count(number);
+        counter(number);
       }
       catch (const std::exception & e)
       {
@@ -51,5 +29,5 @@ int main()
     }
   }
   while (number != 0);
-  std::cout << counter.get_result() << "\n";
+  std::cout << counter() << "\n";
 }
