+++ sakovskaia.maiia/P1/sequence_counter.hpp
@@ -1,0 +1,16 @@
+#ifndef SEQUENCE_COUNTER_HPP
+#define SEQUENCE_COUNTER_HPP
+
+#include <cstddef>
+namespace sakovskaia
+{
+  struct SequenceCounter
+  {
+    SequenceCounter();
+    void operator()(int);
+    size_t operator()() const;
+  private:
+    size_t count_;
+  };
+}
+#endif
